hanging on repo call

+++ acme/src/main/java/com/acme/request/model/DisasterSummaryOutput.java
@@ -5,10 +5,10 @@ import lombok.Data;
 
 @Data
 public class DisasterSummaryOutput {
-    private Integer unaffected;
-    private Integer injured;
-    private Integer illInFacility;
-    private Integer illNotInFacility;
-    private Integer deceased;
-    private Integer isolated;
+    private Long unaffected;
+    private Long injured;
+    private Long illInFacility;
+    private Long illNotInFacility;
+    private Long deceased;
+    private Long isolated;
 }
